Refactor: Migrate financial values from `Double` to `BigDecimal` for improved precision
- **Data Layer:**
    - Updated `MovimentacaoDao` and `MovimentacaoRepository` to return `Flow<BigDecimal?>` for total balance.
    - Modified `MovimentacaoEntity` and `ParcelaEntity` to store `valor` and `porcentagemJuros` as `BigDecimal`.

- **ViewModels:**
    - **`AdicionarContaViewModel.kt`**: Updated interest rate and installment value calculations to use `BigDecimal` arithmetic and `HALF_EVEN` rounding.
    - **`PaginaInicialViewModel.kt`**: Refactored revenue (`receita`) state and `adicionaReceita` logic to handle `BigDecimal` values.

- **UI Components:**
    - **`BreezeCardConta.kt`**: Updated spending and interest logic to support `BigDecimal`.
    - **`ContaPrincipal.kt`**: Refactored `retornaValorTotalArredondado` to perform precise multiplication with `BigDecimal`.
    - **`EditarValorConta.kt`**: Added an optional `modifier` parameter and updated internal state to handle `BigDecimal` conversions.
    - **`ConfirmPaymentModel.kt` & `ParcelaUI.kt`**: Updated `juros` and `valor` fields to `BigDecimal`.

- **Utilities (`Utils.kt`):**
    - Updated `formataSaldo`, `formataValorConta`, and `formataTaxaDeJuros` to accept `BigDecimal` parameters.
    - Refactored `MoneyUtils.arredondarValor` to return `BigDecimal` directly, maintaining `HALF_EVEN` rounding.

+++ app/src/main/java/com/migueldk17/breeze/data/local/dao/MovimentacaoDao.kt
@@ -7,6 +7,7 @@ import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import com.migueldk17.breeze.data.local.entity.MovimentacaoEntity
 import kotlinx.coroutines.flow.Flow
+import java.math.BigDecimal
 
 @Dao
 interface MovimentacaoDao {
@@ -16,7 +17,7 @@ interface MovimentacaoDao {
 
     //Pega a soma de todos os valores em saldo_table
     @Query("SELECT SUM(valor) FROM movimentacao_entity")
-    fun getSaldoTotal(): Flow<Double?>
+    fun getSaldoTotal(): Flow<BigDecimal?>
 
     //Busca o primeiro registro da tabela saldo_table
     @Query("SELECT * FROM movimentacao_entity ORDER BY data DESC")
